Move testcases to velox backend
